Bug#35157: mysqldump should use FLUSH TABLES NO_WRITE_TO_BINLOG \ 	when --master-data is used
When using the --master-data option with mysqldump, mysqldump uses 
a FLUSH TABLES command.  However, this statement got replicated to 
the slave(s), which caused the slave(s) to block unnecessarily while
the FLUSH tables command completed.

Now, if the master-data option is set to one of the two "on" modes,
then use the "LOCAL" qualifier to ensure that it's not replicated.


client/mysqldump.c:
  If master_data is set to one of the two modes, then insert "LOCAL"
  to the command to FLUSH TABLES so that the slaves aren't told to 
  flush also.
mysql-test/r/mysqldump.result:
  Output of mysqldump changed.