Bug#34108 "Drop function" (test "storedproc_06") returns different message (all handlers)
The problem is that the patch for Bug 33464 didn't update the
results of the test cases which caused the problem.

Dropping a stored routine also revoke privileges for all users
on the stored routine and errors about missing grants are converted
into warnings. Before Bug 33464 such errors could be incorrectly
returned to the user, which would later trigger a assertion due to
multiple errors being set.