ndb - wl3600++
- misc cleanup
1) remove unused variables and struct in dict
2) replace the awful c_schemaRecord.schemaPage != 0 with a enum
   and replaced the "ugly" assignment for masterCheckSchemaStatus
   byt a logical memcpy   

- remove backup ongoing
This patch was originally planned to be the grand unified object state-patch
However, lazyness made into simply removing BACKUP_ONGOING and replacing
this with a variable on the table-record.

a step in the right direction, although not grand

- flush schema phases
This patch introduces 3 new transaction phases
flush (prepare/commit/complete)

They are all signal ping-ponged and state checked, but does not actually
do anything serious.