Fix race in ENABLE_COMORD signal.
When QMGR sends ENABLE_COMORD to CMVMI to enable IO with a node
(IOState), it is necessary that this signal is executed before other
node starts sending signals, otherwise signals can get lost.

There was no mechanism to ensure this, it just so happened that
current scheduler ensured this. But it will break with ndbmtd.

Fix by replacing with ENABLE_COMREQ / ENABLE_COMCONF signal pair.