branches/innodb+: buf_page_get_gen(): Correct some ut_ad() assertions.
zip_size should always equal fil_space_get_zip_size(space), even when called
from ibuf.

ibuf_page() should be called with mtr=NULL to be equivalent to the original
implementation.  We cannot assume that mtr holds an x-latch on the insert
buffer bitmap page that covers the page that is being requested.