branches/innodb+: ibuf_rec_get_size(): Correct some bugs.  rb://17
This function was created when the delete buffering code was merged from
branches/fts in r2324.

ibuf_rec_get_size(): Rename the parameter new_format to pre_4_1.  Use
the correct offsets and accessors for both values of pre_4_1.

ibuf_rec_get_info(): Note that the record must be in 4.1 or later format.