branches/zip: Add some more patches to mysql-test.
type_bit_innodb.diff: Ignore the reported duplicate value.  The reported
value depends on the internal workings of the merge sort in
fast index creation.  In the built-in InnoDB of MySQL 5.1, the records
are copied to a new temporary table in the order they were selected from
the original table.

information_schema.diff: Account for the INFORMATION_SCHEMA tables that
were added in the InnoDB plugin.

innodb_file_per_table.diff: In the InnoDB plugin, innodb_file_per_table
is a settable global variable.