branches/zip: for building dynamic plugin on Windows, ha_innodb.dll, when INNODB_DYNAMIC_PLUGIN is specified.
The changes are:

CMakeLists.txt: add project ha_innodb for dynamic plugin on Windows.
ha_innodb depends on project mysqld.

ha_innodb.def: a new file with standard exports for a dynamic plugin.

Two new files will be added:
 * sql/mysqld.def:	.def file for 32-bit compiler
 * sql/mysqld_x64.def:	.def file for x64 compiler

It is also required to apply a patch to the MySQL source tree. The 
patch is described in win-plugin/README:

win-plugin/win-plugin.diff - a patch to be applied to MySQL source
tree. When applied, the following files will be modified:
 * CMakeLists.txt:	add INNODB_DYNAMIC_PLUGIN and _USE_32BIT_TIME_T
 * sql/CMakeLists.txt:	add mysqld.def or mysqld_x64.def for mysqld
 * win/configure.js:	add INNODB_DYNAMIC_PLUGIN
 * win/build-vs71.bat:	provide an option to specify CMAKE_BUILD_TYPE
 * win/build-vs8.bat:	provide an option to specify CMAKE_BUILD_TYPE
 * win/build-vs8_x64.bat: provide an option to specify CMAKE_BUILD_TYPE