BUG#40707 rpl_multi_engine fails sporadically in pushbuild
If before running test rpl_mulit_engin, the mysqltest1 database exists
on master but not exists on slave, then the following statement:

  create database if not exists mysqltest1;

would not be logged to binary log, and so the database would not be
created on slave. This would cause the test to fail and reporting 
mysqltest1 database not existed on slave.

This patch fixed this problem by not using a different database for
the whole test, just use the default test database, there is no reason
why a seperate database is used for this test.


mysql-test/include/rpl_multi_engine.inc:
  use default test database instead of mysqltest1
mysql-test/suite/rpl/r/rpl_multi_engine.result:
  use default test database instead of mysqltest1
mysql-test/suite/rpl/t/rpl_multi_engine.test:
  use default test database instead of mysqltest1
  add missing --enable_warnings