Bug #37742: HA_EXTRA_KEYREAD flag is set when key contains only prefix of requested column
When the storage engine uses secondary keys clustered with the primary key MySQL was
adding the primary key parts to each secondary key.
In doing so it was not checking whether the index was on full columns and this
resulted in the secondary keys being added to the list of covering keys even if 
they have partial columns.
Fixed by not adding a primary key part to the list of columns that can be used 
for index read of the secondary keys when the primary key part is a partial key part.

mysql-test/r/innodb_mysql.result:
  Bug #37742: test case
mysql-test/t/innodb_mysql.test:
  Bug #37742: test case
sql/table.cc:
  Bug #37742: don't add the primary key part to the list of covering key parts
  of a secondary key if it's a partial key part.