WL4258 Long Signal Transactions
Correcting handling of Signal Dropped Rep for fragmented signals.

When the transporter cannot deliver an incoming long signal to a worker thread
due to a shortage of Section Segments, it instead delivers a Signal Dropped
Report.

When the signal is part of a larger fragmented signal, this can result in 
node failures, or inconsistent data being received as any parts which are
correctly received, are incorrectly handled.

This patch adds handling for the case where a dropped signal is part of
a fragmented signal.

Testing is done using SCANTABREQ which can be fragmented as a result of
WL4258.