WL4258 Long Signal Transactions
Previous work in WL4258 modified NdbRecord based PK and UK operations to 
send long signals to the kernel while non-NdbRecord PK and UK operations
continued to send 'short' signals.

This patch modifies NDBAPI to send long signals for non-NdbRecord PK and
UK operations so that after this patch, *all* operations are sent using long signals.

Most of the modifications are limited to the send-phase.